+++ src/cdm/FolderModel.ts
@@ -1,7 +1,7 @@
 import { DatabaseView } from "DatabaseView";
 import { Dispatch } from "react";
 import StateManager from "StateManager";
-import { RowSelectOption } from "cdm/RowSelectModel";
+import { RowSelectOption } from "cdm/ComponentsModel";
 import NoteInfo from "services/NoteInfo";
 import { TFile } from "obsidian";
 import { Column } from "react-table";